requirements.txt
asgiref==3.3.4
attrs==21.2.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
click==8.0.1
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
decorator==5.0.9
dj-database-url==0.5.0
Django==3.2
django-cors-headers==3.7.0
django-heroku==0.3.1
django-rest-knox==4.1.0
djangorestframework==3.12.4
drf-yasg==1.20.0
dynaconf==3.1.4
gunicorn==20.1.0
hvac==0.10.14
idna==2.10
importlib-metadata==4.5.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jsonschema==3.2.0
loguru==0.5.3
MarkupSafe==2.0.1
packaging==20.9
psycopg2==2.9.1
pycparser==2.20
pyparsing==2.4.7
pyrsistent==0.17.3
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.4
shortuuid==1.0.1
simplejson==3.17.2
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.5
validators==0.18.2
whitenoise==5.3.0
win32-setctime==1.0.3
zipp==3.4.1
